Update 01-intro.md
I suggest we remove the parts between "Let's get started" and before "## Nelle's Pipeline: A Typical Problem" and move it to the next episode and put "## Nelle's Pipeline: A Typical Problem" after "We can build on the command-line skills covered here to tackle a wide range of scientific questions and computational challenges." because the part that explains ls seems a bit unrelated to the section before and after it. However, in the next episode, the actual coding in shell starts and would be better matched to the contents removed from here.