Update 02-filedir.md
I suggest we remove the parts between "Let's get started" and before "## Nelle's Pipeline: A Typical Problem" from previous episode and move it to this episode because the part that explains shell and ls is not a perfect fit to the section before and after in previous episode. However, in this episode, the actual working with shell starts and would be better matched to these contents.